build uodates 10-12-20

+++ app/Http/Livewire/Connect/Profile/EditProfile.php
@@ -21,16 +21,16 @@ class EditProfile extends Component
     public $tag;
     public $description;
 
-    public function mount()
-    {
+    public function mount() {
+        $this->authorize('access', $this->profile);
         $this->name = $this->profile->name;
         $this->description = $this->profile->description;
         $this->tag = $this->profile->tag;
     }
 
-    public function update(Request $request)
-    {
-        $this->authorize('update', $this->profile);
+
+    public function update(Request $request) {
+        $this->authorize('access', $this->profile);
 
         $this->validate($this->rules());
 
@@ -48,11 +48,11 @@ class EditProfile extends Component
 
         $this->emitSelf('saved');
 
-        return redirect("/" . User::find($request->user()->id)->profile->data_slug('name') . "/@{$this->profile->tag}");
+        return redirect(route('profile.edit', ['user' => User::find($request->user()->id)->profile->data_slug('name'), 'tag' => $this->profile->tag]));
     }
 
-    public function deleteProfilePhoto()
-    {
+    public function deleteProfilePhoto() {
+        $this->authorize('access', $this->profile);
         return $this->profile->deleteProfilePhoto();
     }
 
@@ -64,10 +64,10 @@ class EditProfile extends Component
                 'min:4',
                 'max:255',
                 ($this->profile->isBusiness() && strtolower($this->profile->name) !== strtolower($this->name)) ?
-                    Rule::unique('profiles', 'name')->where(function ($query) {
-                        return $query->where('profileable_type', 'App\Models\Business');
-                    })
-                    : '',
+                Rule::unique('profiles', 'name')->where(function ($query) {
+                    return $query->where('profileable_type', 'App\Models\Business');
+                })
+                : '',
             ],
 
             'tag' => (strtolower($this->profile->tag) !== strtolower($this->tag)) ? [
@@ -77,11 +77,11 @@ class EditProfile extends Component
                 'unique:profiles,tag',
                 'alpha_dash'
             ] :
-                [
-                    'required',
-                    'min:4',
-                    'alpha_dash'
-                ],
+            [
+                'required',
+                'min:4',
+                'alpha_dash'
+            ],
 
             'description' => [
                 'required',
@@ -96,23 +96,19 @@ class EditProfile extends Component
         ];
     }
 
-    public function getProfileProperty()
-    {
+    public function getProfileProperty() {
         return Profile::find($this->profileId);
     }
 
-    public function updated($propertyName)
-    {
+    public function updated($propertyName) {
         $this->validateOnly($propertyName, $this->rules());
     }
 
-    protected function upperCaseWords(string $string)
-    {
+    protected function upperCaseWords(string $string) {
         return ucwords(strtolower($string));
     }
 
-    public function render()
-    {
+    public function render() {
         return view('livewire.connect.profile.edit-profile', [
             'tag_prefix' => Profile::TAG_PREFIX
         ]);
